[Feat & Chore] 웨이브 시작 UI 추가 및 카메라 관련 기능 수정
- 카메라 영역 추가
- 고래에 카메라 흔들림 추가
- 웨이브 시작을 알려주는 UI 추가

+++ RabbitAndTurtle/Assets/08. Scripts/UI/PlayerUI.cs
@@ -39,7 +39,7 @@ public class PlayerUI : MonoBehaviour
         float shield = _shieldStat.CurrentGauge / _shieldStat.MaxBlockTime;
         _shieldBar.fillAmount = shield;
 
-        float exp = _levelData.Exp / _levelData.MaxExp;
+        float exp = _levelData.Exp / (float)_levelData.MaxExp;
         _expBar.fillAmount = exp;
     }
 }
